C.Parse

Description

Documentation

data ATKind a

Syntax tree type

Constructors

ATAdd

ATSub

ATMul

ATDiv

ATNum a

data ATree a

Abstract syntax tree

Constructors

ATEmpty

ATNode (ATKind a) (ATree a) (ATree a)

inners :: ([Token i] -> ATree i -> Maybe ([Token i], ATree i)) -> [(Char, ATKind i)] -> [Token i] -> ATree i -> Maybe ([Token i], ATree i)

inners is a general function for creating expr and factor in the following syntax (EBNF) of LL (1).

\[ \begin{eqnarray} {\rm expr} &=& {\rm term}\ ("+"\ {\rm term}\ \mid\ "-"\ {\rm term})^\ast \\ {\rm term} &=& {\rm factor}\ ("\ast"\ {\rm factor}\ \mid\ "/"\ {\rm factor})^\ast \\ {\rm factor} &=& {\rm num} \mid\ "(" {\rm expr} ")" \end{eqnarray} \]

expr :: [Token i] -> ATree i -> Maybe ([Token i], ATree i)

expr indicates \({\rm expr} = {\rm term}\ ("+"\ {\rm term}\ \mid\ "-"\ {\rm term})^\ast\) among the comments of inners. This is equivalent to the following code:

expr ::  [Token i] -> ATree i -> Maybe ([Token i], ATree i)
expr xs atn = flip (maybe Nothing) (factor xs atn) $ \(ert, erat) -> expr' ert erat
    where
        expr' (TKReserved '+':ys) era = maybe Nothing (\(zrt, zrat) -> expr' zrt $ ATNode ATAdd era zrat) $ factor ys era
        expr' (TKReserved '-':ys) era = maybe Nothing (\(zrt, zrat) -> expr' zrt $ ATNode ATSub era zrat) $ factor ys era
        expr' ert era = Just (ert, era)

factor :: [Token i] -> ATree i -> Maybe ([Token i], ATree i)

factor indicates \({\rm factor} = {\rm num} \mid\ "(" {\rm expr} ")"\) amount the comments of inners.

term :: [Token i] -> ATree i -> Maybe ([Token i], ATree i)

term indicates \({\rm term} = {\rm factor}\ ("\ast"\ {\rm factor}\ \mid\ "/"\ {\rm factor})^\ast\) amont the comments of inners. Same as expr.

parse :: [Token i] -> Maybe (ATree i)

Constructs the abstract syntax tree based on the list of token strings. if construction fails, Nothing is returned.