Htcc.Asm.Generate

Description

The executable module for compilation

Documentation

type InputCCode = Text

input string, C source code

Generator

casm' :: (Integral e, Show e, Integral i, IsOperand i, UnaryInstruction i, BinaryInstruction i) => ASTs i -> GlobalVars i -> Literals i -> Asm AsmCodeCtx e ()

Executor that receives information about the constructed AST, global variables, and literals and composes assembly code

buildAST :: (Integral i, Read i, Show i, Bits i) => InputCCode -> ASTResult i

Build AST from string of C source code

execAST :: (Integral i, Read i, Show i, Bits i) => Bool -> FilePath -> InputCCode -> IO (Maybe (ASTs i, GlobalVars i, Literals i))

Print warning or error message if building AST from string of C source code has some problems