Htcc.Asm.Intrinsic.Structure.Internal

Description

The modules of intrinsic (x86_64) assembly

Documentation

newtype Asm ctx e a

A monad that represents the context of the assembly code

Constructors

Asm

Fields

unAsm :: AsmInfo e -> IO a

Function that determines the structure of assembly code

data AsmInfo e

Counter and label information used when generating assembly code

Constructors

AsmInfo

Fields

inLabel :: Bool

the flag that indicates whether it is inside the label. If True, indent by single tab,

lblCnt :: IORef e

the label counter

brkCnt :: IORef (Maybe e)

the break label counter

cntCnt :: IORef (Maybe e)

the continue label counter

curFn :: IORef (Maybe Text)

the function being processed

data AsmCodeCtx

Type representing assembly code

unCtx :: Asm ctx e a -> Asm ctx' e a

the function to switch context

runAsm :: (Num e, Enum e) => Asm AsmCodeCtx e a -> IO a

the executor that outputs assembly code

putStrWithIndent :: Text -> Asm ctx e ()

print a string with indentation

putStrLnWithIndent :: Text -> Asm ctx e ()

print a string with indentation, output is broken on a new line

errCtx :: Text -> Asm ctx e ()

The error context. when this is executed, it will exit the application immediately with exitFailure after printing the message.

writeCurFn :: Maybe Text -> Asm ctx e ()

rewriting functions during processing

section :: Text -> Asm ctx e a -> Asm AsmCodeCtx e a

represents a section of assembly code

labeled :: Asm ctx e a -> Asm ctx e a

switch to process in label