Htcc.Asm.Intrinsic.Structure.Section.Data

Description

The modules of intrinsic (x86_64) assembly

Documentation

data DataSectionCtx

the type representing the context inside the data section

data DataLabelCtx

the type representing the context inside the data label

dAta :: Asm DataSectionCtx e a -> Asm AsmCodeCtx e a

data section

label :: Text -> Asm DataLabelCtx e a -> Asm DataSectionCtx e a

label in data section.

byte :: ByteString -> Asm DataLabelCtx e ()

byte in data section

sbyte :: (Num i, Show i) => Natural -> i -> Asm DataLabelCtx e ()

.x.byte in data section

ascii :: ByteString -> Asm DataLabelCtx e ()

ascii in data section

asciiz :: ByteString -> Asm DataLabelCtx e ()

asciiz in data section

zero :: Natural -> Asm DataLabelCtx e ()

zero in data section

quad :: Text -> Asm DataLabelCtx e ()

quad in data section