Htcc.Asm.Intrinsic.Structure.Section.Text.Directive

Description

The modules of intrinsic (x86_64) assembly

Context type

data TextSectionCtx

the type representing the context inside the text section

data TextLabelCtx

the type representing the context inside the label

data TargetLabelCtx

the type representing the context inside the instruction that needs to be specified, such as a jmp instruction.

Directives

text :: Asm TextSectionCtx e a -> Asm AsmCodeCtx e a

text section

global :: Text -> Asm TextSectionCtx e ()

global directive

Labels

fn :: Text -> Asm TextLabelCtx e a -> Asm TextSectionCtx e a

the label as function definition in text section

label :: (Show i, Show e) => Text -> i -> Asm TextLabelCtx e ()

the label in text section

begin :: (Show e, Show i) => i -> Asm TextLabelCtx e ()

begin label

end :: (Show e, Show i) => i -> Asm TextLabelCtx e ()

end label

eLse :: (Show e, Show i) => i -> Asm TextLabelCtx e ()

else label

cAse :: (Show e, Show i) => i -> Asm TextLabelCtx e ()

case label

break :: (Show e, Show i) => i -> Asm TextLabelCtx e ()

break label

continue :: (Show e, Show i) => i -> Asm TextLabelCtx e ()

continue label

gotoLabel :: Text -> Asm TextLabelCtx e ()

goto label

ref :: (Show e, Show i) => Text -> i -> Asm TargetLabelCtx e ()

reference to general label

refBegin :: (Show e, Show i) => i -> Asm TargetLabelCtx e ()

reference to begin label

refEnd :: (Show e, Show i) => i -> Asm TargetLabelCtx e ()

reference to end label

refElse :: (Show e, Show i) => i -> Asm TargetLabelCtx e ()

reference to else label

refBreak :: (Show e, Show i) => i -> Asm TargetLabelCtx e ()

reference for break label

refHBreak :: Show e => Asm TargetLabelCtx e ()

reference for break label (applying value by brkCnt)

refContinue :: (Show e, Show i) => i -> Asm TargetLabelCtx e ()

reference for continue label

refHContinue :: Show e => Asm TargetLabelCtx e ()

reference for break label (applying value by cntCnt)

refReturn :: Show e => Asm TargetLabelCtx e ()

reference for return label

refGoto :: Text -> Asm TargetLabelCtx e ()

reference for goto label

Generator

makeCases :: (Show e, Enum e, Integral e, Show i, Num i) => [ATree i] -> Asm TextLabelCtx e [ATree i]

generate cases and return abstract tree