Htcc.Asm.Intrinsic.Structure.Section.Text.Instruction

Description

The modules of intrinsic (x86_64) assembly

Documentation

data SizeUnit

Unit of size of data to be loaded

Constructors

Byte
8 bits

Word
16 bits

DWord
32 bits

class Show a => UnaryInstruction a where

A class of x86_64 instructions with unary arguments.

Minimal complete definition

Nothing

Methods

push :: a -> Asm TextLabelCtx e ()

push instruction

pop :: a -> Asm TextLabelCtx e ()

pop instruction

pushl :: a -> Asm TextLabelCtx e ()

pushl instruction

popl :: a -> Asm TextLabelCtx e ()

popl instruction

idiv :: a -> Asm TextLabelCtx e ()

idiv instruction

not :: a -> Asm TextLabelCtx e ()

not instruction

class Show a => BinaryInstruction a where

A class of x86_64 instructions with binary arguments.

Minimal complete definition

Nothing

Methods

mov :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

mov instruction

movl :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

movl instruction

movsx :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

movsx instruction

movsxd :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

movsxd instruction

movabs :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

movabs instruction

movzb :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

movzb instruction

cmp :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

cmp instruction

add :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

add instruction

sub :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

sub instruction

imul :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

imul instruction

and :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

and instruction

or :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

or instruction

xor :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

xor instruction

shl :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

shl instruction

sar :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

sar instruction

lea :: BinaryInstruction b => a -> b -> Asm TextLabelCtx e ()

newtype Offset

offset instruction

Constructors

Offset Text
The constructor of offset instruction

data Ptr a

The ptr instruction

Constructors

Ptr SizeUnit (Ref a)
The constructor of ptr instruction

sete :: Register -> Asm TextLabelCtx e ()

sete instruction

setne :: Register -> Asm TextLabelCtx e ()

setne instruction

setl :: Register -> Asm TextLabelCtx e ()

setl instruction

setle :: Register -> Asm TextLabelCtx e ()

setle instruction

setg :: Register -> Asm TextLabelCtx e ()

setg instruction

setge :: Register -> Asm TextLabelCtx e ()

setge instruction

byte :: IsOperand a => (SizeUnit -> Ref a -> Ptr a) -> Ref a -> Ptr a

byte is a helper function for intuitively writing byte instructions

word :: IsOperand a => (SizeUnit -> Ref a -> Ptr a) -> Ref a -> Ptr a

word is a helper function for intuitively writing word instructions

dword :: IsOperand a => (SizeUnit -> Ref a -> Ptr a) -> Ref a -> Ptr a

dword is a helper function for intuitively writing dword instructions

cqo :: Asm TextLabelCtx e ()

cqo instruction

ret :: Asm TextLabelCtx e ()

ret instruction

leave :: Asm TextLabelCtx e ()

leave instruction

jmp :: Asm TargetLabelCtx e () -> Asm TextLabelCtx e ()

jmp instruction

je :: Asm TargetLabelCtx e () -> Asm TextLabelCtx e ()

je instruction

jne :: Asm TargetLabelCtx e () -> Asm TextLabelCtx e ()

jne instruction

jnz :: Asm TargetLabelCtx e () -> Asm TextLabelCtx e ()

jnz instruction

call :: Text -> Asm TextLabelCtx e ()

call instruction