Htcc.CRules.Char

Description

Characters rules of C language

The definition of characters rules

isValidChar :: Char -> Bool

Return True if it is a valid character.

The helper class for some string types

class GenericStr a where

Class of type that can be treated as a set of characters.

Methods

isValid :: a -> Bool

Returns True if the set of characters is a valid C language characters.