Htcc.CRules.Types.TypeKind

Description

The types of C language

TypeKind data type

data StructMember i

The type and offset value of a data member.

Constructors

StructMember
StructMember constructor

Fields

smType :: TypeKind i

The type of a data member

smOffset :: Natural

The offset of a data member

data TypeKind i

The kinds of types in C language.

Constructors

CTInt
The type int as C language

CTChar
The type char as C language

CTSigned (TypeKind i)
The type signed as C language

CTShort (TypeKind i)
The type short as C language

CTLong (TypeKind i)
The type long as C language

CTBool
The type _Bool as C language

CTVoid
The type void as C language

CTPtr (TypeKind i)
The pointer type of TypeKind

CTArray Natural (TypeKind i)
The array type

CTEnum (TypeKind i) (Map Text i)
The enum, has its underlying type and a map

CTStruct (Map Text (StructMember i))
The struct, has its members and their names.

CTIncomplete (Incomplete i)
The incomplete type.

CTUndef
Undefined type

data Incomplete i

The type representing an incomplete type

Constructors

IncompleteArray (TypeKind i)
incomplete array, it has a base type.

IncompleteStruct Text
incomplete struct, it has a tag name.

Type classes that can be converted to TypeKind

class TypeKindBase a where

Class to a type based on TypeKind.

Methods

isCTArray :: a i -> Bool

isCTArray returns True when the given argument is CTArray. Otherwise, returns False

isArray :: a i -> Bool

isArray return True when the given argument is CTArray or IncompleteArray Otherwise, returns False

isCTStruct :: a i -> Bool

isCTStruct returns True when the given argument is CTStruct. Otherwise, returns False

isCTUndef :: a i -> Bool

isCTUndef returns True when the given argument is CTUndef. Otherwise, returns False

isCTIncomplete :: a i -> Bool

isCTIncomplete returns True when the given argument is CTIncomplete.

makeCTArray :: [Natural] -> a i -> a i

makeCTArray retunrs a multidimensional array based on the arguments (list of each dimension). e.g.:

>>> makeCTArray [1, 2] CTInt
int[1][2]
>>> makeCTArray [1, 2] (CTArray 2 CTInt)
int[2][1][2]

concatCTArray :: Ord i => a i -> a i -> Maybe (a i)

Only if both arguments is CTArray, concatCTArray returns a new multidimensional array by conbining the types of multidimensional arrays as follows.

>>> makeCTArray [1, 2] CTInt `concatCTArray` makeCTArray [3, 4] CTInt
Just int[1][2][3][4]
>>> CTInt `concatCTArray` CTArray 2 CTInt
Nothing

toTypeKind :: a i -> TypeKind i

Convert to TypeKind.

mapTypeKind :: (TypeKind i -> TypeKind j) -> a i -> a j

Application to TypeKind.

class IncompleteBase a where

A class requesting a type that represents an incomplete type.

Methods

isIncompleteArray :: a i -> Bool

When the given argument is incomplete array, isIncompleteArray returns True, otherwise False.

isIncompleteStruct :: a i -> Bool

When the given argument is incmoplete struct, isIncompleteStruct returns True, otherwise False.

fromIncompleteStruct :: a i -> Maybe Text

Extract the tag name from IncompleteStruct. If not IncompleteStruct, Nothing is retunred.

fromIncompleteArray :: a i -> Maybe (TypeKind i)

Extract the type of array from IncompleteArray. If not IncompleteArray, Nothing is retunred.

isValidIncomplete :: Ord i => a i -> Bool

Returns True if the incomplete type is temporarily valid at the time of declaration. Otherwise returns False.

Lookup functions

lookupMember :: Text -> TypeKind i -> Maybe (StructMember i)

lookupMember search the specified member by its name from CTStruct.

Utilities of C type

alignas :: (Bits a, Num a, Enum a) => a -> a -> a

alignas align to n.

data Desg i

The type of designator

Constructors

DesgIdx i
index type

DesgMem (StructMember i)
struct member type

accessibleIndices :: Integral i => TypeKind i -> [[Desg i]]

If the given argument is CTArray, it returns a list of accessible indexes of the array. Othrewise returns empty list.