Htcc.Parse

Description

Documentation

data LVar a

The local variable

Constructors

LVar
The constructor of local variable

Fields

name :: String

The name of local variable

offset :: a

The offset value

data ATKind a

The syntax tree type

Constructors

ATAdd
\(+\)

ATSub
\(-\)

ATMul
\(\times \)

ATDiv
\(\div\)

ATLT
\(\lt\)

ATLEQ
\(\leq\)

ATGT
\(\gt\)

ATGEQ
\(\geq\)

ATEQ
\(=\)

ATNEQ
\(\not=\)

ATNum a
The number

ATAssign
The assign operator

ATReturn
The return keyword

ATIf
The if keyword

ATElse
The else keyword

ATWhile
The while keyword

ATLVar a
The local variable. It has a offset value

data ATree a

The data structure of abstract syntax tree

Constructors

ATEmpty
The empty node

ATNode (ATKind a) (ATree a) (ATree a)
ATKind representing the kind of node and the two branches ATree it has

inners :: ([Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])) -> [(String, ATKind i)] -> [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

inners is a general function for creating equality, relational, add and term in the following syntax (EBNF) of LL (1).

\[ \begin{eqnarray} {\rm program} &=& {\rm stmt}^\ast\label{eq:eigth}\tag{1}\\ {\rm stmt} &=& \begin{array}{l} {\rm expr}\ {\rm ";"}\\ \mid\ {\rm "return"}\ {\rm expr}\ ";"\\ \mid\ "{\rm if}"\ "("\ {\rm expr}\ ")"\ {\rm stmt}\ ("{\rm else}"\ {\rm stmt})?\\ \mid\ {\rm "while"\ "("\ expr\ ")"\ stmt}\\ \mid\ {\rm "for"\ "("\ expr?\ ";" expr?\ ";"\ expr?\ ")"\ stmt} \end{array}\label{eq:nineth}\tag{2}\\ {\rm expr} &=& {\rm assign}\\ {\rm assign} &=& {\rm equality} \left("="\ {\rm assign}\right)?\label{eq:seventh}\tag{3}\\ {\rm equality} &=& {\rm relational}\ \left("=="\ {\rm relational}\ \mid\ "!="\ {\rm relational}\right)^\ast\label{eq:fifth}\tag{4}\\ {\rm relational} &=& {\rm add}\ \left("\lt"\ {\rm add}\mid\ "\lt ="\ {\rm add}\mid\ "\gt"\ {\rm add}\mid\ "\gt ="\ {\rm add}\right)^\ast\label{eq:sixth}\tag{5}\\ {\rm add} &=& {\rm term}\ \left("+"\ {\rm term}\ \mid\ "-"\ {\rm term}\right)^\ast\label{eq:first}\tag{6} \\ {\rm term} &=& {\rm factor}\ \left("\ast"\ {\rm factor}\ \mid\ "/"\ {\rm factor}\right)^\ast\label{eq:second}\tag{7} \\ {\rm unary} &=& \left("+"\ \mid\ "-"\right)?\ {\rm factor}\label{eq:fourth}\tag{8} \\ {\rm factor} &=& {\rm num} \mid\ {\rm ident} \mid\ "(" {\rm add} ")"\label{eq:third}\tag{9} \end{eqnarray} \]

add :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

add indicates \(\eqref{eq:first}\) among the comments of inners.

term :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

term indicates \(\eqref{eq:second}\) amont the comments of inners.

unary :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

unary indicates \(\eqref{eq:fourth}\) amount the comments of inners.

factor :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

factor indicates \(\eqref{eq:third}\) amount the comments of inners.

relational :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

relational indicates \(\eqref{eq:sixth}\) among the comments of inners.

equality :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

equality indicates \(\eqref{eq:fifth}\) among the comments of inners. This is equivalent to the following code:

equality ::  [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i)
equality xs atn vars = flip (maybe Nothing) (relational xs atn vars) $ uncurry3 equality'
    where
        equality' (TKReserved "+":ys) era ars = maybe Nothing (uncurry id . first3 equality' . second3 (ATNode ATEQ era)) $ relational ys era ars
        equality' (TKReserved "-":ys) era ars = maybe Nothing (uncurry id . first3 equality' . second3 (ATNode ATNEQ era)) $ relational ys era ars
        equality' ert era ars = Just (ert, era, ars)

assign :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

assign indicates \(\eqref{eq:seventh}\) among the comments of inners.

expr :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

expr is equivalent to equality.

stmt :: Num i => [Token i] -> ATree i -> [LVar i] -> Maybe ([Token i], ATree i, [LVar i])

stmt indicates \(\eqref{eq:nineth}\) among the comments of inners.

NOTE:

Current status

\[ \begin{eqnarray} {\rm stmt} &=& {\rm expr\ ";"}\\ &\mid& {\rm "if"\ "("\ expr\ ")"\ stmt}\ \left({\rm "else"\ stmt}\right)?\\ &\mid& {\rm "while"\ "("\ expr\ ")"\ stmt}\\ &\mid& {\rm "return"\ expr\ ";"} \end{eqnarray} \]

program :: Num i => [Token i] -> [LVar i] -> Maybe [(ATree i, [LVar i])]

program indicates \(\eqref{eq:eigth}\) among the comments of inners.

parse :: Num i => [Token i] -> Maybe ([ATree i], Int)

Constructs the abstract syntax tree based on the list of token strings. if construction fails, Nothing is returned. parse is equivalent to program.