Htcc.Parser.ConstructionData.Scope

Description

The Data type of variables and its utilities used in parsing

The types

data Scoped i

The data type of a struct tag

Constructors

Scoped
The constructor of a struct tag

Fields

curNestDepth :: !Natural

The nest depth of the parsing process

vars :: Vars i

scoped all identifiers of variables (local variables, global variables and literals) visible during processing

structs :: Tags i

scoped all struct tags

typedefs :: Typedefs i

scoped all typedefs

functions :: Functions i

scoped all identifires of functions

enumerators :: Enumerators i

scoped all identifiers of enumerators

data LookupVarResult i

A type that represents the result of a variable search

Constructors

FoundGVar (GVar i)
A type constructor indicating that a global variable has been found

FoundLVar (LVar i)
A type constructor indicating that a local variable has been found

FoundEnum (Enumerator i)
A type constructor indicating that a enumerator has been found

NotFound
A type constructor indicating that it was not found

Operations for scope

addLVar :: (Integral i, Bits i) => StorageClass i -> TokenLC i -> Scoped i -> Either (ASTError i) (ATree i, Scoped i)

addLVar has a scoped type argument and is the same function as addLVar internally.

addGVar :: (Integral i, Bits i) => StorageClass i -> TokenLC i -> Scoped i -> Either (ASTError i) (ATree i, Scoped i)

addGVar has a scoped type argument and is the same function as addGVar internally.

addGVarWith :: (Integral i, Bits i) => StorageClass i -> TokenLC i -> GVarInitWith i -> Scoped i -> Either (ASTError i) (ATree i, Scoped i)

addGVarWith has a scoped type argument and is the same function as addLiteral internally.

addLiteral :: (Integral i, Bits i) => StorageClass i -> TokenLC i -> Scoped i -> Either (ASTError i) (ATree i, Scoped i)

addLiteral has a scoped type argument and is the same function as addLiteral internally.

addTag :: Num i => StorageClass i -> TokenLC i -> Scoped i -> Either (ASTError i) (Scoped i)

addTag has a scoped type argument and is the same function as add internally.

addTypedef :: (Eq i, Num i) => StorageClass i -> TokenLC i -> Scoped i -> Either (ASTError i) (Scoped i)

addTypedef has a scoped type argument and is the same function as add internally.

addFunction :: Num i => Bool -> StorageClass i -> TokenLC i -> Scoped i -> Either (ASTError i) (Scoped i)

addFunction has a scoped type argument and is the same function as add internally.

addEnumerator :: Num i => StorageClass i -> TokenLC i -> i -> Scoped i -> Either (ASTError i) (Scoped i)

addEnumerator has a scoped type argument and is the same function as add internally.

succNest :: Scoped i -> Scoped i

succNest has a scoped type argument and is the same function as succNest internally.

fallBack :: Scoped i -> Scoped i -> Scoped i

fallBack has a scoped type argument and is the same function as fallBack internally.

lookupLVar :: Text -> Scoped i -> Maybe (LVar i)

lookupLVar has a scoped type argument and is the same function as lookupLVar internally.

lookupGVar :: Text -> Scoped i -> Maybe (GVar i)

lookupGVar has a scoped type argument and is the same function as lookupGVar internally.

lookupVar :: Text -> Scoped i -> LookupVarResult i

lookupVar has a scoped type argument and is the same function as lookupVar internally.

lookupTag :: Text -> Scoped i -> Maybe (Tag i)

lookupTag has a scoped type argument and is the same function as lookupTag internally.

lookupTypedef :: Text -> Scoped i -> Maybe (Typedef i)

lookupTypedef has a scoped type argument and is the same function as lookupTypedef internally.

lookupFunction :: Text -> Scoped i -> Maybe (Function i)

lookupFunction has a scoped type argument and is the same function as lookupFunction internally.

lookupEnumerator :: Text -> Scoped i -> Maybe (Enumerator i)

lookupEnumerator has a scoped type argument and is the same function as lookupFunction internally.

initScope :: Scoped i

Helper function representing an empty scoped data

resetLocal :: Scoped i -> Scoped i

resetLocal has a scoped type argument and is the same function as resetLocal internally.