Htcc.Parser.ConstructionData.Scope.Enumerator

Description

The Data type of variables and its utilities used in parsing

Documentation

data Enumerator i

The data type of a enumerator

Constructors

Enumerator

Fields

enVal :: i

The value of enumerator

enUnderlying :: StorageClass i

The underlying type of this enumerator

type Enumerators i = Map Text (Enumerator i)

The typedefs data typedefs

add :: Num i => StorageClass i -> TokenLC i -> i -> Enumerators i -> Either (ASTError i) (Enumerators i)

Given the flag (when that is added function, it is True. otherwise False), type, identifier token, and Enumerators, if the specified identifier already exists in the same scope, return an error message and its location as a pair. Otherwise, add a new tag to Enumerators and return it. If the token does not indicate an identifier, an error indicating internal compiler error is returned.