Htcc.Parser.ConstructionData.Scope.Function

Description

The Data type of variables and its utilities used in parsing

Documentation

data Function a

The data type of a typedef tag

Constructors

Function
The contypedefor of a typedef tag

Fields

fntype :: StorageClass a

The type of this typedef

fnDefined :: Bool

If the function is defined, it will be True, otherwise will be False.

type Functions i = Map Text (Function i)

The typedefs data typedefs

add :: Num i => Bool -> StorageClass i -> TokenLC i -> Functions i -> Either (ASTError i) (Functions i)

Given the flag (when that is added function, it is True. otherwise False), type, identifier token, and Functions, if the specified identifier already exists in the same scope, return an error message and its location as a pair. Otherwise, add a new tag to Functions and return it. If the token does not indicate an identifier, an error indicating internal compiler error is returned.