Htcc.Parser.ConstructionData.Scope.ManagedScope

Description

The Data type of variables and its utilities used in parsing

Documentation

class ManagedScope a where

Type classes common to concepts managed in scope

Methods

lookup :: Text -> Map Text a -> Maybe a

lookup searches for something managed by the scope by the specified Text from a.

fallBack :: Map Text a -> Map Text a -> Map Text a

Organize a list state after scoping.

initial :: Map Text a

Helper function representing an empty something managed by the scope

type ASTError i = (Text, TokenLC i)

The type to be used if an error occurs during AST construction.