Htcc.Parser.ConstructionData.Scope.Typedef

Description

The Data type of variables and its utilities used in parsing

Documentation

data Typedef a

The data type of a typedef tag

Constructors

Typedef
The contypedefor of a typedef tag

Fields

tdtype :: StorageClass a

The type of this typedef

tdNestDepth :: !Natural

The nest depth of this typedef

type Typedefs a = Map Text (Typedef a)

The typedefs data type

add :: (Num i, Eq i) => Natural -> StorageClass i -> TokenLC i -> Typedefs i -> Either (ASTError i) (Typedefs i)

Given the current nesting number, type, identifier token, and Typedefs, if the specified identifier already exists in the same scope, return an error message and its location as a pair. Otherwise, add a new tag to Typedefs and return it. If the token does not indicate an identifier, an error indicating internal compiler error is returned.