Htcc.Parser.ConstructionData.Scope.Var

Description

The Data type of variables and its utilities used in parsing

The classes

class Var a where

The type class of the type representing the variable

Methods

vtype :: a i -> StorageClass i

vtype returns the type of the variable

The data type

type SomeVars v = Map Text v

The type of variables

data GVarInitWith i

The informations type about initial value of the global variable

Constructors

GVarInitWithZero

GVarInitWithOG Text

GVarInitWithVal i

data GVar i

The data type of the global variable

Constructors

GVar
The constructor of the global variable

Fields

gvtype :: StorageClass i

The type of the global variable

initWith :: GVarInitWith i

The informations about initial value of the global variable

data LVar a

The data type of local variable

Constructors

LVar
The constructor of local variable

Fields

lvtype :: StorageClass a

The type of the local variable

rbpOffset :: !a

The offset value from RBP

nestDepth :: !Natural

The nest depth of a local variable

data Literal a

The literal

Constructors

Literal
The literal constructor

Fields

litype :: StorageClass a

The single literal type

ln :: !Natural

The number of labels placed in the .data section

lcts :: ByteString

The content

type GlobalVars a = SomeVars (GVar a)

The type of global variables

type LocalVars a = SomeVars (LVar a)

The type of global variables

type Literals a = [Literal a]

The type of literals

data Vars a

The data type of local variables

Constructors

Vars
The constructor of variables

Fields

globals :: GlobalVars a

The global variables

locals :: LocalVars a

The local variables

literals :: Literals a

Literals

Functions for adding and searching for variables and literals

lookupLVar :: Text -> Vars a -> Maybe (LVar a)

Search for a local variable with a given name

lookupGVar :: Text -> Vars a -> Maybe (GVar a)

Search for a global variable with a given name

lookupVar :: Text -> Vars a -> Maybe (Either (GVar a) (LVar a))

First, search for local variables, and if not found, search for global variables. If nothing is found, Nothing is returned

addLVar :: (Integral i, Bits i) => Natural -> StorageClass i -> TokenLC i -> Vars i -> Either (ASTError i) (ATree i, Vars i)

If the specified token is TKIdent and the local variable does not exist in the list, addLVar adds a new local variable to the list, constructs a pair with the node representing the variable, wraps it in Right and return it. Otherwise, returns an error message and token pair wrapped in Left.

addGVarWith :: Num i => StorageClass i -> TokenLC i -> GVarInitWith i -> Vars i -> Either (ASTError i) (ATree i, Vars i)

If the specified token is TKIdent and the global variable does not exist in the list, addLVar adds a new global variable to the list, constructs a pair with the node representing the variable, wraps it in Right and return it. Otherwise, returns an error message and token pair wrapped in Left.

addGVar :: Num i => StorageClass i -> TokenLC i -> Vars i -> Either (ASTError i) (ATree i, Vars i)

If the specified token is TKIdent and the global variable does not exist in the list, addLVar adds a new global variable that will be initialized by zero to the list, constructs a pair with the node representing the variable, wraps it in Right and return it. Otherwise, returns an error message and token pair wrapped in Left.

addLiteral :: (Ord i, Num i) => StorageClass i -> TokenLC i -> Vars i -> Either (ASTError i) (ATree i, Vars i)

If the specified token is TKString, addLiteral adds a new literal to the list, constructs a pair with the node representing the variable, wraps it in Right and return it. Otherwise, returns an error message and token pair wrapped in Left.

Utilities

initVars :: Vars a

Helper function representing an empty variables

resetLocal :: Vars a -> Vars a

resetLocal initialize the local variable list for Vars

fallBack :: Vars a -> Vars a -> Vars a

Organize variable list state after scoping