Htcc.Token

Description

The tokenizer

Token data types

data Token i

Token type

Constructors

TKReserved String
The reserved token

TKNum i
The number data

TKIdent String
The identifier

TKReturn
The return keyword

TKIf
The if keyword

TKElse
The else keyword

TKWhile
The while keyword

TKFor
The for keyword

TKEmpty
The empty token (This is not used by tokenize, but when errors are detected during parsing, the token at error locations cannot be specified)

type TokenIdx i = (i, Token i)

Token and its index.

Tokenizer

tokenize :: (Integral i, Read i, Show i) => String -> Either (i, Text) [TokenIdx i]

Tokenize the String. If an invalid chraracter matches as C language, the part and the character are returned. Otherwise, [TokenIdx i] is returned.

Helper functions

takeBrace :: forall i. (Integral i, Read i, Show i) => String -> String -> [TokenIdx i] -> Maybe (Either (TokenIdx i) ([TokenIdx i], [TokenIdx i]))

Extract the partial token enclosed in parentheses from the token sequence. If it is invalid, takeBrace returns (i, Text) indicating the error location. Otherwise, takeBrace returns a partial token enclosed in parentheses and subsequent tokens.

takeExps :: Eq i => [TokenIdx i] -> Maybe [[TokenIdx i]]

Get arguments from list of Token (e.g. Given the token of "f(f(g(a, b)), 42);", return expressions that are the token of "f(g(a, b))" and the token of "42".

isTKIdent :: Token i -> Bool

Utility for TKIdent. When the argument is TKIdent, it returns True, otherwise False.

isTKNum :: Token i -> Bool

Utility for TKNum. When the argument is TKNum, it returns True, otherwise False.

isTKReserved :: Token i -> Bool

Utility for TKReserved. When the argument is TKReserved, it returns True, otherwise False.