Htcc.Tokenizer.Token

Description

Types used in lexical analysis and their utility functions

Token data types

data TokenLCNums i

TokenLCNums is data structure for storing the line number and character number of each token

Constructors

TokenLCNums
The constructor of TokenLCNums

Fields

tkLn :: !i

line number

tkCn :: !i

character number

type TokenLC i = (TokenLCNums i, Token i)

Token and its TokenLCNums.

data Token i

Token type

Constructors

TKReserved Text
The reserved token

TKNum !i
The number data

TKIdent Text
The identifier

TKReturn
The return keyword

TKIf
The if keyword

TKSwitch
The switch keyword

TKCase
The case keyword

TKDefault
The default keyword

TKElse
The else keyword

TKWhile
The while keyword

TKFor
The for keyword

TKBreak
The break keyword

TKContinue
The continue keyword

TKEnum
The enum keyword

TKSizeof
The sizeof keyword

TKAlignof
The _Alignof keyword

TKStruct
The struct keyword

TKGoto
THe goto keyword

TKType (StorageClass i)
Types

TKTypedef
The typedef keyword

TKString ByteString
The string literal

TKMacro Macros Text
The C macro

TKEmpty
The empty token (This is not used by tokenize, but when errors are detected during parsing, the token at error locations cannot be specified)

Utilities for accessing to token data

length :: Show i => Token i -> Int

length returns the token length

emptyToken :: Num i => TokenLC i

emptyToken is used when it cannot be referenced

isTKNum :: Token i -> Bool

Utility for TKNum. When the argument is TKNum, it returns True, otherwise False.

isTKType :: Token i -> Bool

Utility for TKType. When the argument is TKType, it returns True, otherwise False.

isTKStruct :: Token i -> Bool

Utility for TKStruct. When the argument is TKStruct, it returns True, otherwise False.

isTKEnum :: Token i -> Bool

Utility for TKEnum. When the argument is TKEnum, it returns True, otherwise False.

isTKIdent :: Token i -> Bool

Utility for TKIdent. When the argument is TKIdent, it returns True, otherwise False.

isTKReserved :: Token i -> Bool

Utility for TKReserved. When the argument is TKReserved, it returns True, otherwise False.

isTKMacro :: Token i -> Bool

Utility for TKMacro. When the argument is TKMacro, it returns True, otherwise False.

isTKString :: Token i -> Bool

Utility for TKString. When the argument is TKString, it returns True, otherwise False.

spanStrLiteral :: Text -> Maybe (Text, Text)

spanStrLiteral separate the string literal part and the non-string literal part from the input text

spanCharLiteral :: Text -> Maybe (Text, Text)

spanCharLiteral separate the string literal part and the non-string literal part from the input text

spanIntLit :: (Eq i, Num i, Read i) => Text -> Maybe (Natural, Token i, Text)

Take the integer literal from given text.

lookupKeyword :: forall i. Show i => Text -> Maybe (Token i)

Lookup keyword from Text. If the specified Text is not keyword as C language, lookupKeyword returns Nothing.

altEmptyToken :: Num i => [TokenLC i] -> TokenLC i

altEmptyToken returns emptyToken if the first token is empty. Otherwise, returns the first token in the token sequence.