Htcc.Utils

Description

General-purpose utilities

Extra functions for lists

safeHead :: [a] -> Maybe a

The exception free implementation of head.

takeWhileLen :: (a -> Bool) -> [a] -> (Int, [a])

takeWhileLen, applied to a predicate f and a list xs, returns the longest prefix (possibly empty) of xs of elements that satisfy f and the length of the list taken. The time complexity of this function is equivalent to takeWhile.

spanLen :: (a -> Bool) -> [a] -> (Int, [a], [a])

Almost the same as span, but returns the number of elements in the list that satisfy f at the same time.

For Data.Text

tshow :: Show a => a -> Text

Convert Show class instance to Text.

For IO shortcuts

putStrErr :: Text -> IO ()

Standard error output shortcut.

putStrLnErr :: Text -> IO ()

Standard error output shortcut (with new line).

err :: Text -> IO ()

Standard error output and exit shortcut.

For triples

first3 :: (a -> d) -> (a, b, c) -> (d, b, c)

Update the first component of triple.

second3 :: (b -> d) -> (a, b, c) -> (a, d, c)

Update the second component of triple.

third3 :: (c -> d) -> (a, b, c) -> (a, b, d)

Update the third component of triple.

dropFst :: (a, b, c) -> (b, c)

Drop first element of triple and returns pair

dropSnd :: (a, b, c) -> (a, c)

Drop second element of triple and returns pair

dropThd :: (a, b, c) -> (a, b)

Drop third element of triple and returns pair

Counter

counter :: Enum a => a -> IO (IO a)

The counter is incremented by one each time it is executed.

Boolean methods

lor :: [a -> Bool] -> a -> Bool

For mappings \(f_i:X\to B\) to an element \(x\in X\) of a set \(X\), \(\displaystyle\bigvee_{i} f_i(x)\) where \(B\) is the boolean domain. This function will stop evaluation when the result of \(f_i(x)\) is True (short circuit evaluation). This is equivalent to:

f1 x || f2 x || f3 x == lor [f1, f2, f3] x

land :: [a -> Bool] -> a -> Bool

For mappings \(f_i:X\to B\) to an element (xin X) of a set \(X\), \(\displaystyle\bigwedge_{i} f_i(x)\) where \(B\) is the boolean domain. This is equivalent to:

f1 x && f2 x && f3 x == land [f1, f2, f3] x

sop :: [a -> Bool] -> [a] -> Bool

Sum of product form. For mappings \(f_i:X\to B\) to an element \(x\in X\) of a set \(X\), \(\displaystyle\bigwedge_{j}\bigvee_{i} f_i(x_j)\) where \(B\) is the Boolean domain. This function will stop evaluation when the result of \(f_i(x)\) is True (short circuit evaluation).

sopText :: [Char -> Bool] -> Text -> Bool

The Text version of sop.

For data type

toInts :: Integral i => i -> [Int]

Convert the instance of Integral to Int. When it cause overflow, express it as a list of Ints divided into multiple values. toInts is useful for functions that have an Int type as an argument. e.g.:

>>> toInts (fromIntegral (maxBound :: Int) + 1 :: Integer)
[9223372036854775807,1]
>>> toInts (fromIntegral (maxBound :: Int) * 3 + 4 :: Integer)
[9223372036854775807,9223372036854775807,9223372036854775807,4]