Htcc.Utils.Print

Description

Utilities of print

Shortcuts of print

putStrErr :: Text -> IO ()

Standard error output shortcut.

putStrLnErr :: Text -> IO ()

Standard error output shortcut (with new line).

err :: Text -> IO ()

Standard error output and exit shortcut.

putDocLn :: Doc -> IO ()

Execute putDoc by applying linebreak to <> at the end of given Doc

putDocErr :: Doc -> IO ()

The shortcut of hPutDoc stderr

putDocLnErr :: Doc -> IO ()

Execute putDocErr by applying linebreak to <> at the end of given Doc

errTxtDoc :: String -> Doc

The Doc used to output an error message (String), it is shortcut of red . text

errCharDoc :: Char -> Doc

The Doc used to output an error message (Char), it is shortcut of red. char

warnTxtDoc :: String -> Doc

The Doc used to output an warning message (String), it is shortcut of magenta . text

warnCharDoc :: Char -> Doc

The Doc used to output an warning message (Char), it is shortcut of magenta . char

locTxtDoc :: String -> Doc

Doc used to output a message (String) about the location, such as the file name and its location, it is shortcut of bold . text

locCharDoc :: Char -> Doc

Doc used to output a message (Char) about the location, such as the file name and its location, it is shortcut of bold . char