Htcc.Parse

Description

The C languge parser and AST constructor

Abstract tree types

data LVar a

The local variable

Constructors

LVar
The constructor of local variable

Fields

name :: String

The name of local variable

offset :: a

The offset value

data ATKindFor a

Specially for syntax tree type

Constructors

ATForkw
The for keyword

ATForInit (ATree a)
The initial section of for statement

ATForCond (ATree a)
The conditional section of for statement

ATForIncr (ATree a)
The incremental section of for statement

ATForStmt (ATree a)
The statement section of for statement

data ATKind a

The syntax tree type

Constructors

ATAdd
\(+\)

ATSub
\(-\)

ATMul
\(\times \)

ATDiv
\(\div\)

ATMod
modulus

ATAnd
bitwise and

ATOr
bitwise or

ATXor
bitwise xor

ATNot
bitwise not

ATShl
left shift \(\lt\lt\)

ATShr
right shift \(\gt\gt\)

ATLT
\(\lt\)

ATLEQ
\(\leq\)

ATGT
\(\gt\)

ATGEQ
\(\geq\)

ATEQ
\(=\)

ATNEQ
\(\not=\)

ATNum a
The number

ATAssign
The assign operator

ATReturn
The return keyword

ATIf
The if keyword

ATElse
The else keyword

ATWhile
The while keyword

ATFor [ATKindFor a]
The for keyword

ATBlock [ATree a]
The block

ATLVar a
The local variable. It has a offset value

ATDefFunc Text (Maybe [ATree a])

ATCallFunc Text (Maybe [ATree a])
The function. It has a offset value and arguments (Maybe).

data ATree a

The data structure of abstract syntax tree

Constructors

ATEmpty
The empty node

ATNode (ATKind a) (ATree a) (ATree a)
ATKind representing the kind of node and the two branches ATree it has

isATForInit :: ATKindFor a -> Bool

An utility of ATForInit. When an argument is ATForInit, return True otherwise False

isATForCond :: ATKindFor a -> Bool

An utility of ATForCond. When an argument is ATForCond, return True otherwise False

isATForStmt :: ATKindFor a -> Bool

An utility ATForStmt. When an argument is ATForStmt, return True otherwise False

isATForIncr :: ATKindFor a -> Bool

An utility ATForIncr. When an argument is ATForIncr, return True otherwise False

fromATKindFor :: ATKindFor a -> ATree a

take ATree data from ATKindFor.

Recursive descent implementation functions

program :: (Show i, Eq i, Num i, Integral i, Read i) => [TokenIdx i] -> [LVar i] -> Either (Text, TokenIdx i) [(ATree i, [LVar i])]

program indicates \(\eqref{eq:eigth}\) among the comments of inners.

stmt :: (Show i, Eq i, Num i, Integral i, Read i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

stmt indicates \(\eqref{eq:nineth}\) among the comments of inners.

inners :: ([TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])) -> [(String, ATKind i)] -> [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

inners is a general function for creating equality, relational, add and term in the following syntax (EBNF) of \({\rm LL}\left(k\right)\) where \(k\in\mathbb{N}\).

\[ \begin{eqnarray} {\rm program} &=& {\rm stmt}^\ast\label{eq:eigth}\tag{1}\\ {\rm stmt} &=& \begin{array}{l} {\rm expr}?\ {\rm ";"}\\ \mid\ {\rm "\{"\ stmt}^\ast\ {\rm "\}"}\\ \mid\ {\rm "return"}\ {\rm expr}\ ";"\\ \mid\ "{\rm if}"\ "("\ {\rm expr}\ ")"\ {\rm stmt}\ ("{\rm else}"\ {\rm stmt})?\\ \mid\ {\rm "while"\ "("\ expr\ ")"\ stmt}\\ \mid\ {\rm "for"\ "("\ expr?\ ";" expr?\ ";"\ expr?\ ")"\ stmt? ";"} \end{array}\label{eq:nineth}\tag{2}\\ {\rm expr} &=& {\rm assign}\\ {\rm assign} &=& {\rm bitwiseOr} \left("="\ {\rm assign}\right)?\label{eq:seventh}\tag{3}\\ {\rm bitwiseOr} &=& {\rm bitwiseXor}\ \left("|"\ {\rm bitwiseXor}\right)^\ast\label{eq:tenth}\tag{4}\\ {\rm bitwiseXor} &=& {\rm bitwiseAnd}\ \left("\hat{}"\ {\rm bitwiseAnd}\right)^\ast\label{eq:eleventh}\tag{5}\\ {\rm bitwiseAnd} &=& {\rm equality}\ \left("\&"\ {\rm equality}\right)^\ast\label{eq:twelveth}\tag{6}\\ {\rm equality} &=& {\rm relational}\ \left("=="\ {\rm relational}\ \mid\ "!="\ {\rm relational}\right)^\ast\label{eq:fifth}\tag{7}\\ {\rm relational} &=& {\rm shift}\ \left("\lt"\ {\rm shift}\mid\ "\lt ="\ {\rm shift}\mid\ "\gt"\ {\rm shift}\mid\ "\gt ="\ {\rm shift}\right)^\ast\label{eq:sixth}\tag{8}\\ {\rm shift} &=& {\rm add}\ \left("\lt\lt"\ {\rm add}\mid\ "\gt\gt"\ {\rm add}\right)^\ast\label{eq:thirteenth}\tag{9}\\ {\rm add} &=& {\rm term}\ \left("+"\ {\rm term}\ \mid\ "-"\ {\rm term}\right)^\ast\label{eq:first}\tag{10} \\ {\rm term} &=& {\rm factor}\ \left("\ast"\ {\rm factor}\ \mid\ "/"\ {\rm factor}\right)^\ast\label{eq:second}\tag{11} \\ {\rm unary} &=& \left("+"\ \mid\ "-"\right)?\ {\rm factor}\mid\ \left("!"\ \mid\ "\sim"\ \right)?\ {\rm unary}\label{eq:fourth}\tag{12} \\ {\rm factor} &=& {\rm num} \mid\ {\rm ident}\ \left({\rm "(" \left(expr\ \left(\left(","\ expr\right)^\ast\right)?\right)? ")"}\right)?\ \mid\ "(" {\rm expr} ")"\label{eq:third}\tag{13} \end{eqnarray} \]

bitwiseOr :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

bitwiseOr indicates \(\eqref{eq:tenth}\) among the comments of inners.

bitwiseXor :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

bitwiseXor indicates \(\eqref{eq:eleventh}\) amont the comments of inners.

bitwiseAnd :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

bitwiseAnd indicates \(\eqref{eq:twelveth}\) among the comments of inners.

add :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

add indicates \(\eqref{eq:first}\) among the comments of inners.

term :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

term indicates \(\eqref{eq:second}\) amont the comments of inners.

unary :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

unary indicates \(\eqref{eq:fourth}\) amount the comments of inners.

factor :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

factor indicates \(\eqref{eq:third}\) amount the comments of inners.

relational :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

relational indicates \(\eqref{eq:sixth}\) among the comments of inners.

equality :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

equality indicates \(\eqref{eq:fifth}\) among the comments of inners. This is equivalent to the following code:

equality ::  [HT.TokenIdx i] -> ATree i -> [LVar i] -> Either (T.Text, HT.TokenIdx i) ([HT.TokenIdx i], ATree i)
equality xs atn vars = flip (either Nothing) (relational xs atn vars) $ uncurry3 equality'
    where
        equality' ((_, HT.TKReserved "+"):ys) era ars = either Left (uncurry id . first3 equality' . second3 (ATNode ATEQ era)) $ relational ys era ars
        equality' ((_, HT.TKReserved "-"):ys) era ars = either Left (uncurry id . first3 equality' . second3 (ATNode ATNEQ era)) $ relational ys era ars
        equality' ert era ars = Right (ert, era, ars)

assign :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

assign indicates \(\eqref{eq:seventh}\) among the comments of inners.

expr :: (Show i, Eq i, Num i) => [TokenIdx i] -> ATree i -> [LVar i] -> Either (Text, TokenIdx i) ([TokenIdx i], ATree i, [LVar i])

expr is equivalent to equality.

Parser

parse :: (Show i, Num i, Eq i, Integral i, Read i) => [TokenIdx i] -> Either (Text, TokenIdx i) ([ATree i], Int)

Constructs the abstract syntax tree based on the list of token strings. if construction fails, Nothing is returned.

varNum :: (Show i, Ord i) => ATree i -> Int

varNum returns the number of variables per function.