Htcc.Utils

Description

General-purpose utilities

Extra functions for lists

module Htcc.Utils.List

For Monad

bothM :: Monad m => (a -> m b) -> (a, a) -> m (b, b)

The monadic both. e.g.:

>>> a <- newIORef (42 :: Int)
>>> b <- newIORef (53 :: Int)
>>> bothM readIORef (a, b) >>= print
(42,53)

(*^*) :: Monad m => (a -> m c) -> (b -> m d) -> (a, b) -> m (c, d) infixr 3

The monadic ***. e.g.:

>>> a <- newIORef 1
>>> b <- newIORef 2
>>> (writeIORef a *^* writeIORef b) (42, 53) >> bothM readIORef (a, b) >>= print
(42,53)

For Data.Maybe

maybe' :: b -> Maybe a -> (a -> b) -> b

maybe` is maybe with changed argument order.

For Char

isStrictSpace :: Char -> Bool

isStrictSpace returns True only if the given string is not a linefeed code and isSpace returns True, otherwise returns False.

For Data.Text

module Htcc.Utils.Text

For Numeric.Natural

toNatural :: Integral i => i -> Natural

toNatural is a shortcut for fromIntegral :: Integral i => i -> Natural

toInteger :: Natural -> Integer

toInteger is a shortcut for fromIntegral :: Natural -> Integer

For print shortcuts

module Htcc.Utils.Print

For triples and quadruple

module Htcc.Utils.Tuple

Boolean methods

module Htcc.Utils.Bool

Natural transformations

module Htcc.Utils.NaturalTransformations

For data type

toInts :: Integral i => i -> [Int]

Convert the instance of Integral to Int. When it cause overflow, express it as a list of Ints divided into multiple values. toInts is useful for functions that have an Int type as an argument. e.g.:

>>> toInts (fromIntegral (maxBound :: Int) + 1 :: Integer)
[9223372036854775807,1]
>>> toInts (fromIntegral (maxBound :: Int) * 3 + 4 :: Integer)
[9223372036854775807,9223372036854775807,9223372036854775807,4]